Sửa refresh listview sau khi thêm, xóa, thêm thông báo xác nhận xóa

+++ Parking lot/QLBaiDoXe/QLBaiDoXe/DSThe.xaml.cs
@@ -31,6 +31,7 @@ namespace QLBaiDoXe
         {
             ThemThe add = new ThemThe();
             add.ShowDialog();
+            ListThe.ItemsSource = Cards.GetAllParkingCards();
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
@@ -46,6 +47,8 @@ namespace QLBaiDoXe
             }
             else
             {
+                if (MessageBox.Show("Bạn có muốn xóa thẻ đã chọn?", "Xác nhận", MessageBoxButton.YesNo) == MessageBoxResult.No)
+                    return;
                 var selectedItems = (dynamic)ListThe.SelectedItems[0];
                 Cards.DeleteCard((long)selectedItems.ParkingCardID);
                 MessageBox.Show("Đã xóa thẻ thành công!");
